+++ src/screens/mainScreen/Market/AuctionsList.js
@@ -1,55 +1,38 @@
 import React from "react";
 import {
-  StyleSheet,
-  Text,
-  View,
-  TouchableOpacity,
-  FlatList,
+	Text,
+	FlatList
 } from "react-native";
 
-import { LinearGradient } from "expo-linear-gradient";
-import PlayerCardOG from "../../../components/PlayerCardOG";
-import PlayerCardMS from "../../../components/PlayerCardMS";
-import PlayerCardMO from "../../../components/PlayerCardMO";
+import OfertasGlobales from "./playerTemplate/OfertasGlobales";
+import MisSubastas from "./playerTemplate/MisSubastas";
+import MisOfertas from "./playerTemplate/MisOfertas";
 
 export default function AuctionsList({
-  auctions = [],
-  opciones = 1,
-  paginate = {},
-  setPage = 0,
-  triggerReload,
+	auctions = [],
+	opciones = 1,
+	paginate = {},
+	setPage = 0,
+	triggerReload,
 }) {
-  return (
-    <>
-      <FlatList
-      removeClippedSubviews={false}
-        showsVerticalScrollIndicator={false}
-        styles={{ width: "100%" }}
-        data={auctions}
-        keyExtractor={(_, index) => index.toString()}
-        ListEmptyComponent={<Text>No se encontraron coincidencias</Text>}
-        ItemSeparatorComponent={() => <View style={styles.separator} />}
-        onEndReached={() => {
-          if (paginate?.page < paginate?.pages-1) {
-            setPage(paginate?.page + 1);
-          }
-        }}
-        renderItem={({ item }) => {
-          if (opciones == 1) return <PlayerCardOG auctionData={item} triggerReload={triggerReload} />;
-          if (opciones == 2) return <PlayerCardMS auctionData={item} />;
-          if (opciones == 3) return <PlayerCardMO auctionData={item} triggerReload={triggerReload}/>;
-        }}
-      />
-    </>
-  );
+	return (
+		<>
+			<FlatList
+				data={auctions}
+				keyExtractor={(_, index) => index.toString()}
+				ListEmptyComponent={<Text>No se encontraron coincidencias</Text>}
+				onEndReached={() => {
+					if (paginate?.page < paginate?.pages - 1) {
+						setPage(paginate?.page + 1);
+					}
+				}}
+				showsVerticalScrollIndicator={false}
+				renderItem={({ item }) => {
+					if (opciones == 1) return <OfertasGlobales auctionData={item} triggerReload={triggerReload} />;
+					if (opciones == 2) return <MisSubastas auctionData={item} />;
+					if (opciones == 3) return <MisOfertas auctionData={item} triggerReload={triggerReload} />;
+				}}
+			/>
+		</>
+	);
 }
-
-const styles = StyleSheet.create({
-  textAñadir: {
-    fontWeight: "bold",
-    fontSize: 12,
-    lineHeight: 18,
-    color: "white",
-    textAlign: "left",
-  },
-});
