+++ src/screens/mainScreen/Market/PlayersList.js
@@ -1,47 +1,32 @@
 import React from "react";
 import {
-  StyleSheet,
-  Text,
-  View,
-  TouchableOpacity,
-  FlatList,
+	Text,
+	FlatList
 } from "react-native";
 
 import { LinearGradient } from "expo-linear-gradient";
-import AddPlayerCard from "../../../components/AddPlayerCard";
+import AddPlayerCard from "./playerTemplate/AddPlayerCard";
 
 export default function PlayersList({
-  players = [],
-  paginate = {},
-  setPage = 0,
-  postAuction
+	players = [],
+	paginate = {},
+	setPage = 0,
+	postAuction
 }) {
-  return (
-    <FlatList
-      showsVerticalScrollIndicator={false}
-      styles={{ width: "100%" }}
-      data={players}
-      keyExtractor={(_, index) => index.toString()}
-      ListEmptyComponent={<Text>No se encontraron coincidencias</Text>}
-      ItemSeparatorComponent={() => <View style={styles.separator} />}
-      onEndReached={() => {
-        if (paginate?.page < paginate?.pages -1) {
-          setPage(paginate?.page + 1);
-        }
-      }}
-      renderItem={({ item }) => {
-        if (!item?.isInLineup) return <AddPlayerCard player={item} postAuction={postAuction}/>;
-      }}
-    />
-  );
+	return (
+		<FlatList
+			data={players}
+			keyExtractor={(_, index) => index.toString()}
+			ListEmptyComponent={<Text>No se encontraron coincidencias</Text>}
+			onEndReached={() => {
+				if (paginate?.page < paginate?.pages - 1) {
+					setPage(paginate?.page + 1);
+				}
+			}}
+			showsVerticalScrollIndicator={false}
+			renderItem={({ item }) => {
+				if (!item?.isInLineup) return <AddPlayerCard player={item} postAuction={postAuction} />;
+			}}
+		/>
+	);
 }
-
-const styles = StyleSheet.create({
-  textAñadir: {
-    fontWeight: "bold",
-    fontSize: 12,
-    lineHeight: 18,
-    color: "white",
-    textAlign: "left",
-  },
-});
